requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.3.3
Deprecated==1.2.14
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.7
importlib-resources==6.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.3
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
pyasn1==0.6.0
pydantic==2.5.2
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.14.5
Pygments==2.17.2
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
pytest-mock==3.12.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
slowapi==0.1.8
sniffio==1.3.0
SQLAlchemy==2.0.30
sqlmodel==0.0.18
starlette==0.37.2
tomli==2.0.1
typer==0.12.3
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.16.0
zipp==3.18.1
